requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pypdf>=4.1.0
pandas>=2.0.0
google-generativeai>=0.4.0
tenacity>=8.2.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
